Refactored core code out
