+++ 4.2/app/index.js
@@ -1,8 +1,8 @@
-import Battery from "./inc/battery";
-import Clock from "./inc/clock";
-import Messaging from "./inc/messaging";
-import UI from "./inc/ui";
-import FileStore from "./inc/fileStore";
+import Battery from "../../core/battery";
+import Clock from "../../core/clock";
+import Messaging from "../../core/messaging";
+import UI from "../../core/ui";
+import FileStore from "../../core/fileStore";
 
 import analytics from "fitbit-google-analytics/app";
