+++ core/common/constants.js
