+++ core/common/utils.js
