+++ core/heartrate.js
