+++ core/steps.js
@@ -1,6 +1,6 @@
 import document from "document";
 import { today, goals } from "user-activity";
-import * as util from "../../common/utils";
+import * as util from "./common/utils";
 
 export default class Steps {
